차량 health 정보 조회, 소모품 health 정보 조회 Vehicle/Consumable 메뉴 Maintenence/Consumable로 이동 소모품 교체 이후 계산 버그 수정

+++ green-fleet/war/app/store/ConsumableHistoryStore.js
@@ -19,7 +19,7 @@ Ext.define('GreenFleet.store.ConsumableHistoryStore', {
 			dateFormat:'time'
 		}, {			
 			name : 'miles_last_repl',
-			type : 'float'
+			type : 'int'
 		}, {
 			name : 'worker',
 			type : 'string'
